lab2/3 int

+++ Lab1/LineSegment.cpp
@@ -2,12 +2,12 @@
 #include "Point2D.h"
 #include "PrimitiveRenderer.h"
 
-void LineSegment::SetA(float x, float y)
+void LineSegment::SetA(int x, int y)
 {
 	this->A.SetX(x);
 	this->A.SetY(y);
 }
-void LineSegment::SetB(float x, float y)
+void LineSegment::SetB(int x, int y)
 {
 	this->B.SetX(x);
 	this->B.SetY(y);
@@ -23,7 +23,7 @@ Point2D LineSegment::GetB()
 }
 
 
-LineSegment::LineSegment(float xa, float ya, float xb, float yb)
+LineSegment::LineSegment(int xa, int ya, int xb, int yb)
 {
 	this->SetA(xa, ya);
 	this->SetB(xb, yb);
