+++ Lab1/PrimitiveRenderer.h
@@ -11,10 +11,10 @@ public:
 	static void drawRectangle(sf::RenderWindow *target);
 	static void drawTringle(sf::RenderWindow *target);
 	static void drawSquare(sf::RenderWindow *target);
-	static void drawLine(sf::RenderWindow* target, float x0, float y0, float x1, float y1);
-	static void drawPoint(sf::RenderWindow *target, float x, float y);
+	static void drawLine(sf::RenderWindow* target, int x0, int y0, int x1, int y1);
+	static void drawPoint(sf::RenderWindow *target, int x, int y);
 	static void drawCustomLine(sf::RenderWindow * target, std::vector<LineSegment> lines);
-	static void drawCustomCircle(sf::RenderWindow *target, float x0, float y0, float r);
-	static void drawCustomElipse(sf::RenderWindow* target, float x0, float y0, float rx, float ry);
+	static void drawCustomCircle(sf::RenderWindow *target, int x0, int y0, int r);
+	static void drawCustomElipse(sf::RenderWindow* target, int x0, int y0, int rx, int ry);
 };
